ejemplo1_SBD/coingecko-mqtt/requirements.txt
paho-mqtt==1.6.1
requests==2.32.3
tenacity==9.0.0